+++ spine-ios/Sources/Spine/Spine.swift
@@ -25,7 +25,7 @@ public class Spine {
 /// when the atlas is no longer in use to release its resources.
 public extension Atlas {
     
-    private static func load(atlasFileName: String, loadFile: (_ name: String) async throws -> Data) async throws -> (Atlas, [Image]) {
+    private static func load(atlasFileName: String, loadFile: (_ name: String) async throws -> Data) async throws -> (Atlas, [CGImage]) {
         let atlasBytes = try await loadFile(atlasFileName)
         guard let atlasData = String(data: atlasBytes, encoding: .utf8) as? NSString else {
             throw "Couldn't read atlas bytes"
@@ -42,7 +42,7 @@ public extension Atlas {
             throw "Couldn't load atlas: \(message)"
         }
         
-        var atlasPages = [Image]()
+        var atlasPages = [CGImage]()
         let numImagePaths = spine_atlas_get_num_image_paths(atlas);
         
         for i in 0..<numImagePaths {
@@ -51,10 +51,10 @@ public extension Atlas {
             }
             let atlasPageFile = String(cString: atlasPageFilePointer)
             let imageData = try await loadFile(atlasPageFile)
-            guard let image = UIImage(data: imageData) else {
+            guard let image = UIImage(data: imageData)?.cgImage else {
                 continue
             }
-            atlasPages.append(Image(uiImage: image))
+            atlasPages.append(image)
         }
         
         // TODO: Paint in Swift?
@@ -65,9 +65,9 @@ public extension Atlas {
     /// Loads an [Atlas] from the file [atlasFileName] in the main bundle or the optionally provided [bundle].
     ///
     /// Throws an [Exception] in case the atlas could not be loaded.
-    public static func fromAsset(_ atlasFileName: String, bundle: Bundle = .main) async throws -> (Atlas, [Image]) {
+    public static func fromAsset(_ atlasFileName: String, bundle: Bundle = .main) async throws -> (Atlas, [CGImage]) {
         return try await Self.load(atlasFileName: atlasFileName) { name in
-            return try bundle.loadAsData(fileName: atlasFileName)
+            return try bundle.loadAsData(fileName: name)
         }
     }
     
@@ -165,6 +165,114 @@ public extension SkeletonData {
     /// Throws an [Exception] in case the skeleton data could not be loaded.
     public static func fromHttp(atlas: Atlas, skeletonURL: URL) throws -> SkeletonData {
         throw "Not implemented"
+    }
+}
+
+public final class SkeletonDrawableWrapper {
+    
+    public let atlas: Atlas
+    public let atlasPages: [CGImage]
+    public let skeletonData: SkeletonData
+    
+    public let skeletonDrawable: SkeletonDrawable
+    public let skeleton: Skeleton
+    public let animationStateData: AnimationStateData
+    public let animationState: AnimationState
+    
+    internal var disposed = false
+    
+    public init(atlas: Atlas, atlasPages: [CGImage], skeletonData: SkeletonData) throws {
+        self.atlas = atlas
+        self.atlasPages = atlasPages
+        self.skeletonData = skeletonData
+            
+        guard let nativeSkeletonDrawable = spine_skeleton_drawable_create(skeletonData.wrappee) else {
+            throw "Could not load native skeleton drawable"
+        }
+        skeletonDrawable = SkeletonDrawable(nativeSkeletonDrawable)
+        
+        guard let nativeSkeleton = spine_skeleton_drawable_get_skeleton(skeletonDrawable.wrappee) else {
+            throw "Could not load native skeleton"
+        }
+        skeleton = Skeleton(nativeSkeleton)
+        
+        guard let nativeAnimationStateData = spine_skeleton_drawable_get_animation_state_data(skeletonDrawable.wrappee) else {
+            throw "Could not load native animation state data"
+        }
+        animationStateData = AnimationStateData(nativeAnimationStateData)
+        
+        guard let nativeAnimationState = spine_skeleton_drawable_get_animation_state(skeletonDrawable.wrappee) else {
+            throw "Could not load native animation state"
+        }
+        animationState = AnimationState(nativeAnimationState)
+    }
+    
+    /// Updates the [AnimationState] using the [delta] time given in seconds, applies the
+    /// animation state to the [Skeleton] and updates the world transforms of the skeleton
+    /// to calculate its current pose.
+    public func update(delta: Float) {
+        if disposed { return }
+        
+        animationState.update(delta: delta)
+        animationState.apply(skeleton: skeleton)
+        
+        skeleton.update(delta: delta)
+        skeleton.updateWorldTransform(physics: SPINE_PHYSICS_UPDATE)
+    }
+    
+    public func dispose() {
+        if disposed { return }
+        disposed = true
+        
+        atlas.dispose()
+        skeletonData.dispose()
+        
+        skeletonDrawable.dispose()
+    }
+}
+
+public extension SkeletonDrawable {
+    
+    public func render() -> [RenderCommand] {
+        var commands = [RenderCommand]()
+        if disposed { return commands }
+        
+        var nativeCmd = spine_skeleton_drawable_render(wrappee)
+        repeat {
+            if let ncmd = nativeCmd {
+                commands.append(RenderCommand(ncmd))
+                nativeCmd = spine_render_command_get_next(ncmd)
+            } else {
+                nativeCmd = nil
+            }
+        } while (nativeCmd != nil)
+        
+        return commands
+    }
+}
+
+public extension RenderCommand {
+    
+    internal var numVertices: Int {
+        Int(spine_render_command_get_num_indices(wrappee))
+    }
+    
+    public var positions: [Float] {
+        let num = numVertices * 2
+        let ptr = spine_render_command_get_positions(wrappee)
+        return (0..<num).compactMap { ptr?[$0] }
+    }
+    
+    public var uvs: [Float] {
+        let num = numVertices * 2
+        let ptr = spine_render_command_get_uvs(wrappee)
+        return (0..<num).compactMap { ptr?[$0] }
+    }
+    
+    public var colors: [Int32] {
+        let num = numVertices
+        let ptr = spine_render_command_get_colors(wrappee)
+        return (0..<num).compactMap { ptr?[$0] }
     }
 }