+++ spine-ios/CHANGELOG.md
@@ -7,3 +7,4 @@
 - Spine C++ Swift Wrapper (#3)
 - Load `Atlas` & `SkeletonData` (#4)
 - Generate Swift classes from `spine-cpp-lite.h` (#5)
+- Draw `SkeletonData` render commands (#6)